AssayResultsForSamplesButton.test.tsx - Convert from spec test file
